requirements.txt
asgiref==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cron_descriptor==2.0.5
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
environ==1.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
PyMySQL==1.1.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
requests==2.32.5
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.11.0
